move test read-creation function to TestUtil
